#32: Adding common API calls for Gemini and Ollama
